Refact Tes sum

+++ src/test/java/com/calculator/CalculatorTest.java
@@ -1,13 +1,16 @@
 package test.java.com.calculator;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
+
 import org.junit.jupiter.api.Test;
+
+import main.java.com.calculator.Calculator;
 
 public class CalculatorTest {
     @Test
     public void testSum() {
         Calculator calculator = new Calculator();
-        int result = calculator.sum(2, 3);
+        double result = calculator.sum(2, 3);
         assertEquals(5, result);
     }
 }
